adjustments

+++ procedury.sql
@@ -3,8 +3,8 @@ CREATE PROCEDURE Wstaw_klienta (
         @haslo VARCHAR(64),
         @imie VARCHAR(64),
 		@nazwisko VARCHAR(64),
-		@nip VARCHAR(64),
-		@nazwa_firmy VARCHAR(64))
+		@nip VARCHAR(64)=NULL,
+		@nazwa_firmy VARCHAR(64)=NULL
 AS
 
 INSERT INTO klient (login, haslo, imie, nazwisko, nip, nazwa_firmy)
@@ -17,11 +17,34 @@ CREATE PROCEDURE Usun_klienta (
 		@login VARCHAR(64)
 		)
 AS
-IF NOT EXISTS (SELECT klient_login 
+BEGIN
+	IF NOT EXISTS (SELECT klient_login 
                                                 FROM zamowienie
                                                 WHERE klient_login = @login)
-BEGIN
-	DELETE FROM klient
-	WHERE login = @login
+
+		DELETE FROM klient
+		WHERE login = @login
+	ELSE
+		PRINT N'Nie mozna usunac podanego klienta'
 END
 GO
+
+CREATE PROCEDURE Wstaw_adres (
+        @login VARCHAR(64),
+        @ulica VARCHAR(64),
+        @nr_domu INT=NULL,
+		@nr_lokalu  INT,
+		@kod_pocztowy VARCHAR(64),
+		@miasto VARCHAR(64))
+AS
+
+INSERT INTO adres (klient_login, ulica, nr_domu, nr_lokalu, kod_pocztowy, miasto)
+        VALUES (@login, @ulica, @nr_domu, @nr_lokalu, @kod_pocztowy, @miasto)
+
+GO
+
+
+
+
+
+
